added coloredlogs and optimized batched reading from disk
with getitem we schedule the read tasks sequentially, even though they happen in parallel this is not efficient. With getitems it paralellizes the read task scheduling and its faster. Still not optimized when some stuff is in memory it is reading items even though it shouldn't. Looking in to it more