refactored the get_pipeline stuff
removed the get_pipeline stuff from the implemented steps. it was weird having a get pipeline method in a pipeline object.
One thing is that if we do that the steps names in the pipeline must be different than the arguments of the constructor of that pipeline. So if a is an arg of the init of a pipeline in the steps you can not name a step 'a'.