implemented the scaler block
The scaler block will take an sklearn (dask implementation of it) scaler object and fit the scaler on the data. The data needs to be reshaped/flattened so thats why we have a block, it handles all of that. This way we can also just change the scaler if we want. It does not cache the scaled data so it will have to recalculate the scaler everytime but if everyting is put to memory you only need 1 computation so its since the data all fits in memory for everyone