the data is properly distributed in not mem
before to mem length was used to move to mem length items from both train and test. So even though we have 5635 items with split 0.2 if you set to mem length to 4508 that also put everything in memory. This was now it will put 80% of to mem length items from train and the rest from test with split 0.2