made first pretrained architecture train using pipeline
This is so convenient. You just make an architecture with forward, instantiate it with hydra and it works.
I made a unet model that uses the pretrained Unet from smp. I do the padding in the arhitecture because that is easier than making it a transformation step. It does also seem to learn with a sigmoid activation.